修正系统页面属性

+++ caesar-backend/admin/src/main/java/com/caesar/controller/TeamgroupManagerController.java
@@ -1,8 +1,11 @@
 package com.caesar.controller;
 
 import com.caesar.entity.CaesarTeamGroup;
+import com.caesar.entity.CaesarUserGroup;
+import com.caesar.entity.vo.CaesarTeamGroupVo;
 import com.caesar.model.JsonResponse;
 import com.caesar.service.TeamGroupService;
+import com.caesar.service.UserGroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,12 +18,14 @@ public class TeamgroupManagerController {
     @Autowired
     TeamGroupService teamGroupService;
 
+    @Autowired
+    UserGroupService userGroupService;
 
 
     @GetMapping("getTeamList")
-    public JsonResponse<List<CaesarTeamGroup>> getTeamList(){
+    public JsonResponse<List<CaesarTeamGroupVo>> getTeamList(){
 
-        List<CaesarTeamGroup> list = teamGroupService.list();
+        List<CaesarTeamGroupVo> list = teamGroupService.getTeamList();
 
         return JsonResponse.success(list);
 
@@ -37,6 +42,10 @@ public class TeamgroupManagerController {
 
     @GetMapping("/deleteTeamGroup")
     public JsonResponse<Boolean> deleteTeamGroup(@RequestParam int id){
+        List<CaesarUserGroup> userGroups = userGroupService.getUserGroupsByGroupId(id);
+        if(null != userGroups && userGroups.size()>0){
+            return JsonResponse.fail("该组存在用户，不能删除.");
+        }
         return JsonResponse.success(teamGroupService.deleteTeamGroup(id));
     }
